requirements.txt
asgiref==3.7.2
# black==24.3.0
click==8.1.7
Django==4.2.10
django-bootstrap5==22.2
django-debug-toolbar==4.3.0
# flake8==7.0.0
gunicorn==20.1.0
# isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg==3.1.18
pycodestyle==2.11.1
pyflakes==3.2.0
pymodbus==3.6.9
python-dotenv==0.19.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
cryptography==44.0.2
django-pam==2.1.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10